feat: Update AdminController to use Trainer model for trainers
This commit updates the `AdminController.js` file to use the `Trainer` model instead of the `User` model for handling trainers. The code changes include importing the `Trainer` model and updating the queries to find trainers based on the new model. This change ensures that the correct model is used for managing trainers in the application.

//TODO: bugfix updateTrainers and deactivateTrainers