feat: Update user routes and controllers for trainers
This commit updates the user routes and controllers in the `user.js` file to include functionality for trainers. The code changes include adding new routes for creating, updating, and deleting trainers, as well as retrieving a list of all trainers. The `AdminController` is updated to use the `Trainer` model instead of the `User` model for managing trainers. These changes ensure that trainers can be properly authenticated and authorized within the application.